Add files via upload
Generate a random whole number between 0 and 15. If the result is divisible by 3 alert "fizz". If the number is divisible by 5, alert "buzz", otherwise print the number to the console.

If a number is divisible by 3 AND 5, alert "fizzbuzz." If the random number is 0, it should not return fizz, buzz or fizzbuzz.